Fix meta tag for readability.
When pasting https://donationsfrom.tech/ into an App that unfurls a vCard, like Slack, content in the <meta description=""/> tag is shown as a byline or chyron. In this context, the lack of capitalization makes the sentence read oddly – "us politicians" rather than "U.S.". 

Changed this to better reflect the intention of the website.